Исправление ошибок интеграции в общую систему.
