+++ FacultativeCourseWork1/Program.cs
@@ -23,6 +23,7 @@ namespace FacultativeCourseWork1
                     "7 - Домашняя работа \"Расчет индекса массы тела\" (выполнил: Самигуллин А.И.)",
                     "8 - Домашняя работа \"Расчет индекса массы тела\" (выполнил: Кулик А.А.)",
                     "9 - Домашняя работа \"Работа с дробями\" (выполнил: Кулик А.А.)",
+                    "10 - Домашняя работа \"Работа с двумерными массивами\" (выполнил: Кулик А.А.)"
                 };
                 Console.Clear();
                 PrintVariants(variants);
@@ -59,6 +60,9 @@ namespace FacultativeCourseWork1
                         break;
                     case 9:
                         HomeWorkKulik.Lesson03Task3_Divides.Run();
+                        break;
+                    case 10:
+                        HomeWorkKulik.Lesson04Task5_TwoDimArrays.Run();
                         break;
                     default:
                         Console.WriteLine("Нет домашней работы под таким номером!");
